refactor(cli): extract shared O_NOFOLLOW secure-file helper
The symlink-safe openSync(O_NOFOLLOW)+writeSync/readSync boilerplate was
duplicated across saveCredentials, loadCredentials, and the export write.
Extract writeSecretFile/readSecretFile into cli/src/lib/secure-file.ts; each
call site is now one line. Behavior unchanged.
